this is about as close as I can get without building auth and model classes

+++ t/07_test_against_net-imap-server.t
@@ -1,31 +1,58 @@
 
 use Test;
 use Net::TCP;
-plan tests => my $tests = 1;
+use Net::IMAP::Simple;
 
-SHHH: {
+plan tests => my $tests = 3;
+
+sub run_tests() {
+    my $imap = Net::IMAP::Simple->new('localhost:7000') or die "connect failed: $Net::IMAP::Simple::errstr";
+
+    ok( not $imap->login(qw(bad login)) );
+    ok( $imap->errstr, qr/disabled/ );
+
+    # run tests here
+    ok(1);
+}
+
+# test support:
+
+my $res = do {
     # NOTE: the imap server emits various startup warnings on import
     local $SIG{__WARN__} = sub {};
-    unless( eval "use Coro; use EV; use Net::IMAP::Server; 1" ) {
-        warn " Net::IMAP::Server not found, skipping all meaningful tests\n";
-        skip(1,1,1) for 1 .. $tests;
-        exit 0;
-    }
+    eval "use Coro; use EV; use Net::IMAP::Server; 1";
+};
+
+unless( $res ) {
+    warn "Net::IMAP::Server not found, skipping all meaningful tests\n";
+    skip(1,1,1) for 1 .. $tests;
+    exit 0;
 }
 
 if( my $pid = fork ) {
     my $imapfh;
-    my $retries = 15;
+    my $retries = 7;
     sleep 1 while (--$retries)>0 and not $imapfh = Net::TCP->new(localhost=>7000);
-    warn " imap server is up: " . <$imapfh> . "\n";
+
+    eval q &
+        END {
+          # warn " murdering imap server (if necessary)\n";
+            kill 15, $pid;
+            waitpid $pid, 0;
+        }1;
+    & or die $@;
+
+    if( not $imapfh ) {
+        warn "unable to start Net::IMAP::Server, skipping all meaningful tests\n";
+        skip(1,1,1) for 1 .. $tests;
+        exit 0;
+    } 
+
+    warn "imap server is up: " . <$imapfh>;
     close $imapfh;
 
-    # run tests here
-    ok(1);
+    run_tests();
 
-    # murder the imap server
-    kill 15, $pid;
-    waitpid $pid, 0;
     exit 0;
 }
 
@@ -35,6 +62,9 @@ open STDERR, ">>informal-imap-server-dump.log";
 open STDOUT, ">>informal-imap-server-dump.log";
 # (we don't really care if the above fails...)
 
+$SIG{ALRM} = sub { exit 0 };
+alarm 30; # this server lasts at most 30 seconds, except perhaps on windows (??)
+
 Net::IMAP::Server->new(
     port        => 7000,
     ssl_port    => 8000,
@@ -43,4 +73,3 @@ Net::IMAP::Server->new(
   # user        => "nobody",
   # group       => "nobody",
 )->run;
-
